refactor: moved authBg image to the features/auth directory
