+++ app/frontend/src/features/auth/resources/auth-bg.png
